feat: add spectrograms, time-series, extended MIR, procedures, and monitor
Analysis upgrade: spectrograms (mel + chroma PNGs), per-beat time-series,
extended MIR (onsets, HPSS, spectral contrast, tonnetz, chord estimation).
Audio loaded once via full_analysis() coordinator. Adds matplotlib dep.

New modules: procedures.py (probe_track, sweep_parameter, mix_check) for
automated multi-step workflows; monitor.py for continuous background
capture+analyze loop.

CLI: -s/-t/-e flags on listen/analyze, new commands (spectrogram, probe,
sweep, mix-check, monitor start/latest). CLAUDE.md composition guide with
command quick-reference, music theory, analysis interpretation, and recipes.
Compositions folder for per-song notes.

Also includes: osc.py fire_scene fix, set_scene_name, capture_bars type
hint fix, and composition scripts from prior session.

+++ lib/capture.py
@@ -53,7 +53,7 @@ def capture_seconds(duration: float) -> Path:
     return outfile
 
 
-def capture_bars(bars: int, bpm: float | None = None) -> Path:
+def capture_bars(bars: float, bpm: float | None = None) -> Path:
     """Capture N bars of audio. If bpm is None, queries Carabiner."""
     if bpm is None:
         from . import link
